CCM-1230:
Refactors test dirs to fix pytest duplicate filename issue.

Every dir needs an __init__.py file to namespace it as a
module and prevent duplicate filenames when pytest collects
and flattens all the tests to run.

Looks like dir names with hyphens in were still raising
duplicate filename errors despite the __init__.py file, so
these have been replaced with underscores.

Renames specification `yml` files to `yaml` so sonar picks them
up.

Adds local SonarCloud scan files.

Adds static analysis to github build actions.
