+++ .gitignore
@@ -30,4 +30,5 @@ zap-report.json
 zap-report.xml
 zap-report.html
 tmp/
+.scannerwork
 .zap
