+++ example.env
@@ -7,3 +7,6 @@ export API_ENVIRONMENT=internal-dev
 # them with the repository
 export INTEGRATION_COLLECTION_API_KEY=xxx
 export SANDBOX_COLLECTION_API_KEY=xxx
+
+# this is required for running the static analysis steps
+export SONAR_TOKEN=xxx
