+++ package.json
@@ -8,6 +8,7 @@
     "serve": "redocly preview-docs -p 5000 build/communications-manager.json",
     "check-licenses": "node_modules/.bin/license-checker --failOn GPL --failOn LGPL --failOn AGPL",
     "sandbox-postman-collection": "newman run postman/CommunicationsManager.Sandbox.postman_collection.json",
+    "static-analysis": "./scripts/perform-static-analysis.sh",
     "zap-security-scan": "./scripts/run_zap.sh"
   },
   "author": "NHS Digital",
