+++ scripts/config/sonar-scanner.properties
@@ -1,0 +1,12 @@
+sonar.host.url=https://sonarcloud.io
+sonar.organization=nhsdigital
+sonar.projectKey=nhsdigital_communications-manager
+sonar.sourceEncoding=UTF-8
+sonar.language=js,python
+
+sonar.exclusions=.venv/**,proxies/utils/performance/*
+
+#exclude everything from test coverage, this is covered by other tools
+sonar.coverage.exclusions=**/*
+sonar.nodejs.executable=/usr/bin/node
+sonar.python.version=3.10.8
