+++ specification/schemas/LinksError.yaml
