+++ specification/schemas/Message.yaml
@@ -7,7 +7,7 @@ properties:
     format: uuid
     example: 703b8008-545d-4a04-bb90-1f2946ce1575
   recipient:
-    $ref: Recipient.yml
+    $ref: Recipient.yaml
   personalisation:
     type: object
     description: The personalisation keys and values for this message. These are linked to the routingPlanId provided and are agreed upon during [onboarding](#overview--onboarding).
