+++ specification/schemas/ServiceTimeoutResponse.yaml
@@ -11,11 +11,11 @@ properties:
       type: object
       properties:
         id:
-          $ref: TypeErrorIdentifier.yml
+          $ref: TypeErrorIdentifier.yaml
         code:
-          $ref: enums/EnumErrorTimeout.yml
+          $ref: enums/EnumErrorTimeout.yaml
         links:
-          $ref: LinksError.yml
+          $ref: LinksError.yaml
         status:
           type: string
           enum:
