+++ specification/schemas/TypeKSUID.yaml
