+++ specification/schemas/enums/EnumErrorAccessDenied.yaml
