+++ specification/schemas/enums/EnumErrorForbidden.yaml
