+++ specification/schemas/enums/EnumErrorInternalServerError.yaml
