+++ specification/schemas/enums/EnumErrorInvalidValue.yaml
