+++ specification/schemas/enums/EnumErrorNoSuchRoutingPlan.yaml
