+++ specification/schemas/enums/EnumErrorNotAcceptable.yaml
