+++ specification/schemas/enums/EnumErrorQuota.yaml
