+++ specification/schemas/enums/EnumErrorServiceUnavailable.yaml
