+++ specification/schemas/enums/EnumErrorUnsupportedMedia.yaml
