+++ specification/schemas/enums/EnumTypeMessageBatch.yaml
