--- tests/development/content-types/test_content_type_negotiation.py
@@ -1,36 +1,0 @@
-import requests
-import pytest
-
-ACCEPT_HEADERS = [
-    {
-        "headers": {
-            "Accept": "application/vnd.api+json"
-        },
-        "expect": "application/vnd.api+json"
-    },
-    {
-        "headers": {
-            "Accept": "*/*"
-        },
-        "expect": "application/vnd.api+json"
-    },
-    {
-        "headers": {
-            "Accept": "application/json"
-        },
-        "expect": "application/json"
-    }
-]
-METHODS = ["get", "post", "put", "patch", "delete", "head", "options"]
-
-
-@pytest.mark.devtest
-@pytest.mark.parametrize("accept_headers", ACCEPT_HEADERS)
-@pytest.mark.parametrize("method", METHODS)
-@pytest.mark.nhsd_apim_authorization({"access": "application", "level": "level3"})
-def test_application_response_type(nhsd_apim_proxy_url, accept_headers, method, nhsd_apim_auth_headers):
-    resp = getattr(requests, method)(f"{nhsd_apim_proxy_url}", headers={
-        **nhsd_apim_auth_headers,
-        **accept_headers.get("headers")
-    })
-    assert resp.headers.get("Content-Type") == accept_headers.get("expect")
